Refactor: move post data validation in usp_check_for_public_submission()
In the end, one of the goals is to make usp_create_public_submission almost only callable with valid arguments. The main goal is to make it throw an exception for invalid arguments (or for any other kind of messages) so that it returns void in the end.
Store classes automatically and permanently verifying the validity of the post data would be a nice addition.
usp_create_public_submission's multiple parameters should be turned into one single object, such as Post_Submission. This would make adding and ignoring fields easier.
Custom exceptions need to be created, and they need to be displayed correctly!